% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_file.R
\name{js_file}
\alias{js_file}
\alias{js_handler}
\alias{css_file}
\title{JS, CSS templates}
\usage{
js_file(path, ...)

js_handler(path, ...)

css_file(path, ...)
}
\arguments{
\item{path, ...}{Args passed from golem}
}
\description{
JS, CSS templates
}
