% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_hook.R
\name{golem_hook}
\alias{golem_hook}
\title{A golem hook example}
\usage{
golem_hook(path, package_name, ...)
}
\arguments{
\item{path, package_name, ...}{args passed to golem_hook}
}
\description{
This hook will automatically add the MIT licence to the project.
}
\examples{
if (interactive()) {
  golem::create_golem("ici", project_hook = golem_hook)
}
}
