% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_template.R
\name{module_template}
\alias{module_template}
\title{A module template}
\usage{
module_template(name, path, ...)
}
\arguments{
\item{name, path, ...}{Arg passed from golem::add_module}
}
\value{

}
\description{
This template will add a module skeleton.
}
\examples{
if (interactive()){
  golem::add_module(
      name = "custom",
      module_template = module_template
  )
}
}
